Get UI talking to new endpoints.
Since we will not be supporting the flattened granule endpoint, we need the UI to use the new collection and granule endpoints to work correctly.

This has been tested manually for collections, using the 490 branch for WIP on staging->search pipelines. Since granule pipelines are not quite complete, this is still pending.