Bug fix: wrong granule search on 'show more'
sometimes the show more on the granule tab caused the wrong granule search to fire - it was caused by a logic error which resulted in clear selections getting called when visiting a detail view after changing the query.

reproducing the bug:
search and select a collection with a granule view, then search with a different query and click a different collection with a granule view.