Create a KTable of parsed attributes and include them in REST responses.
This involved declaring the topics that we need directly because apparently more complicated topologies only work with pre-declared topics. So I created a MetadataStreamService which creates and manages the streams app, including declaring topics up front. Also, there are new combined_* topics which is a join of the raw and parsed information for each key.