First pass at allowing script to return add'l attrs, like publish flag
Scripts can now return xml directly, pure json, or json with an isoXml attribute containing xml. If xml is present it will be parsed, and in the case of xml within json the attributes parsed from the xml will be merged into the json attributes.

Not positive this is the right thing to do, but it's a first step that's working.