Simplify and consolidate kubernetes definitions.
Uses a stateful set with volume claims to preserve elasticsearch data. Uses initContainers to set the vm.max_map_count setting. Added a headless service for elasticsearch node discovery, so we can scale it up or down now. Client is a load balancer service, which works both on minikube and GKE.