Updated promote and snapshot strategy
Migrating from travis to circle for builds revealed a squirrelly dependency on the build system for tagging snapshots.

That dependency is removed, and it is expected that the actual version committed in the code will include -SNAPSHOT during development, until a release is tagged.

The promote.sh script has been updated to help manage that. It allows arguements as before, but if none are provided will attempt to prompt you to provide the desired information.