508 check on presets select:
Everything technically works, but it's not the greatest user experience. It's very easy to change the selection and keep reloading the page just trying to read what the options are. I considered if there were different controls that might improve this, but I can't think of anything that isn't overly cumbersome. (The recommendations are not to auto-submit changes to select objects).

I've decided to leave it as an auto-submitting select for lack of better options overall. It is a shorthand option to assist and guide users with suggestions, but is not at all required. Additionally, it has only a handful of options and submitting multiple times as you scroll through, while not ideal, is at worse overly verbose.

It would be useful to collect actual user data to see how often this feature is used.